requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
easyocr==1.7.1
filelock==3.13.1
fsspec==2023.12.2
idna==3.6
imageio==2.33.1
Jinja2==3.1.2
lazy_loader==0.3
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.2.1
ninja==1.11.1.1
numpy==1.26.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
packaging==23.2
Pillow==10.1.0
pyclipper==1.3.0.post5
python-bidi==0.4.2
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
scikit-image==0.22.0
scipy==1.11.4
shapely==2.0.2
six==1.16.0
sympy==1.12
tifffile==2023.12.9
torch==2.1.2
torchvision==0.16.2
triton==2.1.0
typing_extensions==4.9.0
urllib3==2.1.0
